Refactored the AlertBox component further.
