+++ src/components/DropdownColumn.jsx
@@ -1,25 +1,28 @@
 import React, { useState } from 'react'
 import {
-    Menu, MenuItem, Button, ListItemIcon, ListItemText, Grid, Snackbar
+    Menu, MenuItem, Button, ListItemIcon, ListItemText, Grid
 } from '@material-ui/core'
-import Alert from '@material-ui/lab/Alert'
+import AlertBox from './AlertBox'
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz'
 import Delete from '@material-ui/icons/Delete'
 import { useMutation, useApolloClient } from '@apollo/client'
 import { DELETE_COLUMN } from '../graphql/column/columnQueries'
 import { COLUMNORDER } from '../graphql/fragments'
-import { boardPageStyles } from '../styles/styles'
 
 const DropdownColumn = ({ columnId, boardId }) => {
-    const [deleteColumn] = useMutation(DELETE_COLUMN)
-    const [anchorEl, setAnchorEl] = useState(null)
     const client = useApolloClient()
-    const classes = boardPageStyles()
+    const [anchorEl, setAnchorEl] = useState(null)
     const [open, setOpen] = useState(false)
-    const snackbarMsg = `This action will permanently remove the selected column from your project and can't be later examined! Are you sure you want to delete it?`
+    const alertMsg = `This action will permanently remove the selected column and the tasks inside the column from your board and can't be later examined! Are you sure you want to delete it?`
+    const [deleteColumn] = useMutation(DELETE_COLUMN)
 
     const handleClick = (event) => {
         setAnchorEl(event.currentTarget)
+    }
+
+    const openSnackbar = () => {
+        setOpen(true)
+        setAnchorEl(null)
     }
 
     const deleteColumnById = () => {
@@ -49,20 +52,6 @@ const DropdownColumn = ({ columnId, boardId }) => {
         client.cache.evict({ id: idToBeDeleted })
     }
 
-    const openSnackbar = () => {
-        setOpen(true)
-        setAnchorEl(null)
-    }
-
-    const handleClose = (option) => {
-        if (option === 'DELETE') {
-            deleteColumnById()
-            deleteColumnFromCache()
-        } else {
-            setOpen(false)
-        }
-    }
-
     return (
         <Grid item>
             <Button
@@ -90,27 +79,14 @@ const DropdownColumn = ({ columnId, boardId }) => {
                     <ListItemText primary="Remove" />
                 </MenuItem>
             </Menu>
-            <Snackbar
-                classes={{ root: classes.snackbar }}
+            <AlertBox
                 open={open}
-                anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
-            >
-                <Alert variant="outlined" severity="error">
-                    <Grid item container direction="column">
-                        <Grid item>
-                            <span id="snackbarMessage">{snackbarMsg}</span>
-                        </Grid>
-                        <Grid item container direction="row" justify="flex-end">
-                            <Button variant="contained" onClick={() => handleClose('UNDO')}>
-                                UNDO
-                            </Button>
-                            <Button color="secondary" variant="contained" onClick={() => handleClose('DELETE')} classes={{ root: classes.snackbarButtonDelete }}>
-                                DELETE
-                            </Button>
-                        </Grid>
-                    </Grid>
-                </Alert>
-            </Snackbar>
+                setOpen={setOpen}
+                alertMsg={alertMsg}
+                deleteColumnById={deleteColumnById}
+                deleteColumnFromCache={deleteColumnFromCache}
+                type={'COLUMN'}
+            />
         </Grid>
     )
 }
