+++ src/styles.css
@@ -25,6 +25,6 @@ html, body {
     font-size: 1.2rem;
     width: 100%;
 }
-#snackbarMessage {
+#alertMessage {
     font-size: 1.3rem;
 }
